Don't expect Pathauto to delete patterns on bundle deletion
The Drupal 7 version of Pathauto had logic to evaluate a bundle
being deleted and see if there were matching patterns that had
to be removed as well. This is something that we don't see a
real need for. Instead, we will let the site administrator
to clean up patterns manually after removing a bundle.